requirements.txt
Cerberus==1.3.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.18
pkg-resources==0.0.0
protobuf==3.10.0
pycparser==2.19
pyOpenSSL==19.0.0
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
